Factor out common playback management functionality into a container mixin
